tampoco

+++ exercises/P21_LargerThanOrEqualTo.php
@@ -10,18 +10,18 @@ class P21_LargerThanOrEqualTo
 
 
         // Get input from the user
-        $numberOne = trim(fgets($GLOBALS['STDIN'] ?? STDIN));
+        $numberOne = (int)trim(fgets($GLOBALS['STDIN'] ?? STDIN));
         
         // Prompt the user for input
         echo "Give the second number:\n";
         
         // Get input from the user
-        $numbertwo = trim(fgets($GLOBALS['STDIN'] ?? STDIN));
+        $numbertwo = (int)trim(fgets($GLOBALS['STDIN'] ?? STDIN));
 
         // Check year value
-        if((int)$numberOne > (int)$numbertwo){
+        if($numberOne > $numbertwo){
             echo "Greater number is: ";
-        }elseif((int)$numberOne < (int)$numbertwo){
+        }elseif($numberOne < $numbertwo){
             echo "Greater number is: ";
         } else{
             echo "The numbers are equal!\n";
